requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
aiohttp==3.6.2
altgraph==0.17
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
future==0.18.2
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==1.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
macholib==1.14
Markdown==3.2.2
matplotlib==3.3.2
multidict==4.7.6
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminaltables==3.1.0
torch==1.5.1
torchvision==0.6.1
tqdm==4.48.0
typing-extensions==3.7.4.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
